Abstract predictors and refactor logReg to predict just for a day
All the new predictors will live in the bin/predictor directory and
should implement train, test, and computerError and writeToFile methods.
Instead of predicting for a whole season, logReg now takes in a game day
schedule and predicts for the games in that day.

To run the application, execute "sh kick-game.sh" or run "MoneyBallMain.py".
The bash script is there for setting up scheduling the script to run
everyday.