Modified table merging and LogRegression code to use only three tables
For 2016-2017 season basketball-reference.com only released the data
of three tables(team stats, opponent stats and miscellaneous). Hence
to comply with the current situation, we modified the table merging and
LogRegression code to use only these three tables. Also generated new
full tables for 2014-2015 and 2015-2016 seasons with the new spec.