Use selenium to retrieve the web page.
The problem with urllib or request libraries are that they cannot
load the page entirely when downloading the page. Selenium can get
around this by initiating a web browser instance and the page source
loaded is what we see in the web browser.

However, we need to find a way to solve the problem of sometimes
waiting too long for the page to load.