Refactor predictor and crawler logic.
Because it is frequently the case that the training and test data
don't have the same dimension, it is often a pain to manually select
the features. This change makes sure that the training data and test
data both use the largest intersection of features.

The chrome webdriver we used is not stable in the time used for
retrieving the web page. This change adds the limit for waiting for
the web page to complete loading. It also downloads the page source
if it timeouts as long as the tables we care about are in the page.